requirements.txt
scrapy
redis
pymongo
bs4
scrapy_redis
requests
lxml
html5lib
Flask
twisted
PyMySQL
tld
chardet
jieba
Pillow
cssselect
beautifulsoup
nltk
goose-extractor
numpy
scipy
pandas
scikit-learn==0.18.1
xgboost